backend/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
pandas
numpy
joblib
openai
python-dotenv
